# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Soccer Players API",
    "description": "Tiny RESTful API using Node.js: allows you to manage soccer players.",
    "contact": {
        "name": "Mauro Bonfietti",
        "url": "https://github.com/maurobonfietti"
    },
    "version": "1.0"
  },
  "paths": {
...
    "/v1/players/{playerId}": {
      "get": {
        "tags": [
          "Players"
        ],
        "summary": "Get one player",
        "description": "Get one player.",
        "operationId": "Getoneplayer",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "description": "The Player ID",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {
              "Access-Control-Allow-Origin": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "*"
                  }
                }
              },
              "Content-Length": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "67"
                  }
                }
              },
              "ETag": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "W/\"43-VN/yW16slSmftfDmPi6L+bOOnrY\""
                  }
                }
              },
              "Date": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Thu, 26 Dec 2019 02:45:56 GMT"
                  }
                }
              },
              "Connection": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "keep-alive"
                  }
                }
              }
            },
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Getoneplayer"
                  }
                },
                "example": [
                  {
                    "myposition": "forward",
                    "fullname": "Lionel Messi",
                    "playerId": "2"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false
      },
      "patch": {
        "tags": [
          "Players"
        ],
        "summary": "Update player",
        "description": "Update a player.",
        "operationId": "Updateplayer",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "description": "The Player ID",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateplayerRequest"
              },
              "example": {
                "myposition": "forward",
                "fullname": "{{$randomFullName}}"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {
              "Access-Control-Allow-Origin": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "*"
                  }
                }
              },
              "Content-Length": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "29"
                  }
                }
              },
              "ETag": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "W/\"1d-0TK7dvCtIV2ofSeBvK5eVoxYrtI\""
                  }
                }
              },
              "Date": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Thu, 26 Dec 2019 02:31:17 GMT"
                  }
                }
              },
              "Connection": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "keep-alive"
                  }
                }
              }
            },
            "content": {
              "application/json; charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/Updateplayer"
                },
                "example": {
                  "message": "Player updated."
                }
              }
            }
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "Players"
        ],
        "summary": "Delete player",
        "description": "Delete a player.",
        "operationId": "Deleteplayer",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "description": "The Player ID",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "",
            "headers": {
              "Access-Control-Allow-Origin": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "*"
                  }
                }
              },
              "ETag": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "W/\"2-vyGp6PvFo4RvsFtPoIWeCReyIC8\""
                  }
                }
              },
              "Date": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "Thu, 26 Dec 2019 02:46:55 GMT"
                  }
                }
              },
              "Connection": {
                "content": {
                  "text/plain": {
                    "schema": {
                      "type": "string"
                    },
                    "example": "keep-alive"
                  }
                }
              }
            },
            "content": {
              "text/plain": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Getoneplayer": {
        "title": "Getoneplayer",
        "required": [
          "myposition",
          "fullname",
          "playerId"
        ],
        "type": "object",
        "properties": {
          "myposition": {
            "type": "string"
          },
          "fullname": {
            "type": "string"
          },
          "playerId": {
            "type": "string"
          }
        },
        "example": {
          "myposition": "forward",
          "fullname": "Lionel Messi",
          "playerId": "2"
        }
      },
...
      "UpdateplayerRequest": {
        "title": "UpdateplayerRequest",
        "required": [
          "myposition",
          "fullname"
        ],
        "type": "object",
        "properties": {
          "myposition": {
            "type": "string"
          },
          "fullname": {
            "type": "string"
          }
        },
        "example": {
          "myposition": "forward",
          "fullname": "{{$randomFullName}}"
        }
      },
      "Updateplayer": {
        "title": "Updateplayer",
        "required": [
          "message"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "example": {
          "message": "Player updated."
        }
      },
...
    }
  },
...
}